user操作

+++ app/api/resources/user.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from ..models.user_info import UserModel
 from ..schema.user_sha import user_args_valid
 from ..utils.format import res
@@ -17,12 +17,17 @@ class User(Resource):
 
     @jwt_required()
     def delete(self, user_id):
-        try:
-            UserModel.delete_by_user_id(user_id)
-            UserLoginModel.delete_by_user_id(user_id)
-            return res(message='User deleted successfully!')
-        except Exception as e:
-            return res(success=False, message="Error: {}".format(e), code=500)
+        jwt_data = get_jwt()
+        role = jwt_data['role']
+        if role == 'admin':
+            try:
+                UserModel.delete_by_user_id(user_id)
+                UserLoginModel.delete_by_user_id(user_id)
+                return res(message='User deleted successfully!')
+            except Exception as e:
+                return res(success=False, message="Error: {}".format(e), code=500)
+        else:
+            return res(success=False, message="Role must be admin")
 
     @jwt_required()
     def put(self, user_id):
@@ -49,20 +54,28 @@ class User(Resource):
 class UserList(Resource):
     @jwt_required()
     def get(self):
-        user_info_list = UserModel.find_all()
-        result = []
-        for user_info in user_info_list:
-            result.append(user_info.dict())
-
-        return res(data=result)
+        jwt_data = get_jwt()
+        role = jwt_data['role']
+        if role == 'admin':
+            user_info_list = UserModel.find_all()
+            result = []
+            for user_info in user_info_list:
+                result.append(user_info.dict())
+            return res(data=result)
+        else:
+            return res(success=False, message="Role must be admin", code=403)
 
 
 class UserByUsername(Resource):
     @jwt_required()
     def get(self, username):
-        user_info = UserModel.find_by_username(username)
-        print(username)
-        if user_info:
-            return user_info.dict()
+        jwt_data = get_jwt()
+        role = jwt_data['role']
+        if role == 'admin':
+            user_info = UserModel.find_by_username(username)
+            if user_info:
+                return user_info.dict()
+            else:
+                return res(message="User not found", code=404)
         else:
-            return res(message="User not found", code=404)
+            return res(success=False, message='Access denied.', code=403)
